first pass at hdlc encoding/decoding fix where a header byte is NOT expected. tests obviously broken - commit before addressing tests

+++ HDLC/PPP.h
@@ -19,9 +19,10 @@ typedef struct ppp_buffer_t
     size_t length;
 } ppp_buffer_t;
 
-int PPP_stuff(ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * stuffed_buffer);
-int PPP_unstuff( ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * stuffed_buffer);
-int parse_PPP_stream(uint8_t new_byte, ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * input_buffer);
+
+size_t PPP_stuff(ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * stuffed_buffer);
+size_t PPP_unstuff( ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * stuffed_buffer);
+size_t parse_PPP_stream(uint8_t new_byte, ppp_buffer_t * unstuffed_buffer, ppp_buffer_t * input_buffer);
 
 #ifdef __cplusplus
 }
